+++ doc/package.nix
@@ -3,7 +3,7 @@
   nixosOptionsDoc,
   runCommand,
   wire-small-dev,
-  wire-dignostics-md,
+  wire-diagnostics-md,
   nix,
   nodejs,
   pnpm,
@@ -56,7 +56,7 @@ stdenv.mkDerivation (finalAttrs: {
   };
   patchPhase = ''
     cat ${optionsDoc} >> ./reference/module.md
-    cat ${wire-dignostics-md} >> ./reference/errors.md
+    cat ${wire-diagnostics-md} >> ./reference/errors.md
     wire inspect --markdown-help > ./reference/cli.md
   '';
   buildPhase = "pnpm run build > build.log 2>&1";