+++ wire/cli/default.nix
@@ -81,7 +81,7 @@
           paths = [ self'.packages.wire-unwrapped-perf ];
         };
 
-        wire-dignostics-md = self'.packages.wire-unwrapped.overrideAttrs {
+        wire-diagnostics-md = self'.packages.wire-unwrapped.overrideAttrs {
           DIAGNOSTICS_MD_OUTPUT = "/build/source";
           installPhase = ''
             mv /build/source/DIAGNOSTICS.md $out