+++ wire/cli/src/cli.rs
@@ -137,7 +137,7 @@ pub struct ApplyArgs {
 
     /// List of literal node names, a literal `-`, or `@` prefixed tags.
     ///
-    /// `-` will read additional values from stdin, seperated by whitespace.
+    /// `-` will read additional values from stdin, separated by whitespace.
     /// Any `-` implies `--non-interactive`.
     #[arg(short, long, value_name = "NODE | @TAG | `-`", num_args = 1..)]
     pub on: Vec<ApplyTarget>,